+++ spec/integration/resource/api200/firmware_bundle/create_spec.rb
@@ -13,14 +13,9 @@ require 'spec_helper'
 
 klass = OneviewSDK::FirmwareBundle
 RSpec.describe klass, integration: true, type: CREATE, sequence: seq(klass) do
-  include_context 'integration context'
+  let(:current_client) { $client }
+  let(:bundle_path) { $secrets['bundle_path'] }
+  let(:spp_path) { $secrets['spp_path'] }
 
-  let(:bundle_path) { 'spec/support/hp-firmware-a1b08f8a6b-HPGH-1.1.i386.rpm' }
-
-  describe '#self.upload' do
-    it 'Upload hotfix' do
-      item = OneviewSDK::FirmwareBundle.add($client, bundle_path)
-      expect(item['uri']).to be
-    end
-  end
+  include_examples 'FirmwareBundleCreateExample', 'integration context'
 end