+++ spec/unit/resource/api300/synergy/firmware_bundle_spec.rb
@@ -11,13 +11,13 @@ RSpec.describe OneviewSDK::API300::Synergy::FirmwareBundle do
     it 'returns a FirmwareDriver resource' do
       options = { 'header' => { 'uploadfilename' => 'file.tar' } }
       expect(@client_300).to receive(:upload_file).with('file.tar', '/rest/firmware-bundles', options, OneviewSDK::Rest::READ_TIMEOUT).and_return({})
-      expect(described_class.add(@client_300, 'file.tar').class).to eq(OneviewSDK::FirmwareDriver)
+      expect(described_class.add(@client_300, 'file.tar').class).to eq(OneviewSDK::API300::Synergy::FirmwareDriver)
     end
 
     it 'should call client.upload_file correctly' do
       options = { 'header' => { 'uploadfilename' => 'file.tar' } }
       expect(@client_300).to receive(:upload_file).with('file.tar', '/rest/firmware-bundles', options, 100).and_return({})
-      expect(described_class.add(@client_300, 'file.tar', 100).class).to eq(OneviewSDK::FirmwareDriver)
+      expect(described_class.add(@client_300, 'file.tar', 100).class).to eq(OneviewSDK::API300::Synergy::FirmwareDriver)
     end
   end
 end